Minor integer promotion issue in a test

+++ src/transport/tests/TestPeerConnections.cpp
@@ -97,7 +97,7 @@ void TestBasicFunctionality(nlTestSuite * inSuite, void * inContext)
     // This define guard will be needed when migrating SecureSessionTable to ObjectPool
     //#if !CHIP_SYSTEM_CONFIG_POOL_USE_HEAP
     // If not using a heap, we can fill the SecureSessionTable
-    for (uint16_t i = 2; i < CHIP_CONFIG_SECURE_SESSION_POOL_SIZE; ++i)
+    for (int i = 2; i < CHIP_CONFIG_SECURE_SESSION_POOL_SIZE; ++i)
     {
         sessions[i] =
             connections.CreateNewSecureSessionForTest(SecureSession::Type::kCASE, static_cast<uint16_t>(i + 6u), kLocalNodeId,
